libnfs: make continue_int 64 bit to avoid sign extension bug for truncate()
Make the continue_int field we use through the internal callbacks a uint64_t
instead of an int.
This fixes a bug for the truncate function where we pass the offset to truncate
to via this field.

The bug is that otherwise we first truncate the field to an int and then
in the callback we cast this int back to a uint64_t again.

If the user called truncate with an offset that is >= 2^31
then :
IF the 1<<31 bit is cleared then we would truncate to (offset & 0xffffffff)
IF the 1<<31 bit is set, we would instead truncate to (offset | 0xffffffff00000000)

Reported-by: doktorstick at github

+++ lib/libnfs.c
@@ -197,7 +197,7 @@ struct nfs_cb_data {
        continue_func continue_cb;
        void *continue_data;
        void (*free_continue_data)(void *);
-       int continue_int;
+       uint64_t continue_int;
 
        struct nfs_fh3 fh;
 
@@ -1552,7 +1552,7 @@ static void nfs_lookup_path_getattr_cb(struct rpc_context *rpc, int status, void
 	nfs_lookup_path_async_internal(nfs, attr, data, &nfs->rootfh);
 }
 
-static int nfs_lookuppath_async(struct nfs_context *nfs, const char *path, int no_follow, nfs_cb cb, void *private_data, continue_func continue_cb, void *continue_data, void (*free_continue_data)(void *), int continue_int)
+static int nfs_lookuppath_async(struct nfs_context *nfs, const char *path, int no_follow, nfs_cb cb, void *private_data, continue_func continue_cb, void *continue_data, void (*free_continue_data)(void *), uint64_t continue_int)
 {
 	struct nfs_cb_data *data;
 	struct GETATTR3args args;
